Add sbt build
