+++ .scalafmt.conf
@@ -1,0 +1,9 @@
+style = defaultWithAlign
+
+align.tokens               = ["%", "%%"]
+danglingParentheses        = true
+docstrings                 = JavaDoc
+indentOperator             = spray
+maxColumn                  = 120
+rewrite.rules              = [RedundantBraces, RedundantParens, SortImports]
+unindentTopLevelOperators = true
