+++ project/plugins.sbt
@@ -1,0 +1,9 @@
+addSbtPlugin("com.dwijnand"      % "sbt-dynver"         % "3.1.0")
+addSbtPlugin("de.heikoseeberger" % "sbt-header"         % "5.0.0")
+addSbtPlugin("com.lightbend.sbt" % "sbt-java-formatter" % "0.4.4")
+addSbtPlugin("com.geirsson"      % "sbt-scalafmt"       % "1.5.1")
+addSbtPlugin("org.xerial.sbt"    % "sbt-sonatype"       % "2.5")
+addSbtPlugin("com.jsuereth"      % "sbt-pgp"            % "1.1.0")
+
+resolvers += Resolver.jcenterRepo
+addSbtPlugin("net.aichler" % "sbt-jupiter-interface" % "0.8.1")
