--- scripts/build.bat
@@ -1,5 +1,0 @@
-javac -d ../build ../src/*.java
-
-cd ../build
-jar cvfm ../scripts/PolynomialFactorization.jar ../scripts/manifest *.class
-cd ../scripts
