--- scripts/manifest_test
@@ -1,3 +1,0 @@
-Manifest-Version: 1.2
-Main-Class: Test
-Created-By: 1.4 (Sun Microsystems Inc.)
