--- scripts/run.bat
@@ -1,1 +1,0 @@
-start java -jar PolynomialFactorization.jar
